requirements.txt
alembic==1.9.4
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
passlib==1.7.4
Pillow==9.4.0
psycopg2-binary==2.9.5
pydantic==1.10.5
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2023.3
requests==2.28.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
